Finished exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -35,8 +35,17 @@ blogsRouter.put('/blogs/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/blogs/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  const blog = await Blog.findById(request.params.id)
+  console.log(blog)
+  console.log(decodedToken.id.toString())
+  if (blog.user.toString() == decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'A blog can only be deleted by the user who added it' })
+  }
 })
 
 module.exports = blogsRouter
